able to upload a multi part file
DB changed to LONGBLOB for EntityParts which is not ideal but we may be able to enforce constraints to support 16mb limit.
I had to switch to longblob because we have to pass a MEDIUMBLOB amount of data encoded as hex to make our insert work with MySQL connector. Even though all the file parts are 16MB it still needs to be longblob to support the output of UNHEX. It should not stay longblob